+++ src/helix-ui/elements/hx-radio/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXRadioElement.html';
-import shadowStyles from './HXRadioElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 /**
  * Applies Shadow DOM to the `<hx-radio>` facade element.
