+++ src/helix-ui/elements/index.js
@@ -24,9 +24,9 @@ export { HXModalElement } from './HXModalElement';
 export { HXPillElement } from './HXPillElement';
 export { HXPopoverElement } from './hx-popover/index.js';
 export { HXProgressElement } from './hx-progress/index.js';
-export { HXRadioControlElement } from './HXRadioControlElement';
-export { HXRadioElement } from './HXRadioElement';
-export { HXRadioSetElement } from './HXRadioSetElement';
+export { HXRadioControlElement } from './hx-radio-control/index.js';
+export { HXRadioElement } from './hx-radio/index.js';
+export { HXRadioSetElement } from './hx-radio-set/index.js';
 export { HXRevealElement } from './hx-reveal/index.js';
 export { HXSearchAssistanceElement } from './hx-search-assistance/index.js';
 export { HXSearchElement } from './HXSearchElement';
